+++ tests/Integration/Modules/Environment.php
@@ -61,7 +61,6 @@ class Environment
     {
         $this->shopId = $shopId;
         Registry::getConfig()->setShopId($shopId);
-        $this->loadShopParameters();
     }
 
     /**
@@ -121,9 +120,6 @@ class Environment
      */
     public function clean()
     {
-        $config = Registry::getConfig();
-        $config->setConfigParam('aModules', null);
-
         $database = \OxidEsales\Eshop\Core\DatabaseProvider::getDb();
         $database->execute("DELETE FROM `oxconfig` WHERE `oxmodule` LIKE 'module:%' OR `oxvarname` LIKE '%Module%'");
         $database->execute('TRUNCATE `oxconfigdisplay`');
@@ -214,19 +210,6 @@ class Environment
         $modules = array_diff(scandir($this->getPathToTestDataDirectory() . 'modules'), array('..', '.'));
 
         return $modules;
-    }
-
-    /**
-     * Loads config parameters from DB and sets to config.
-     */
-    protected function loadShopParameters()
-    {
-        $aParameters = array(
-            'aModules'
-        );
-        foreach ($aParameters as $sParameter) {
-            Registry::getConfig()->setConfigParam($sParameter, $this->getConfigValueFromDB($sParameter));
-        }
     }
 
     /**