+++ CHANGELOG.md
@@ -22,6 +22,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 - Switched to templating-engine agnostic names in Controller templates (e.g. `Controller::$_sThisTemplate = 'page/content'` instead of `'page/content.tpl'`)
 - Don't store module controllers in database
 - Store information about active modules state in the module configuration (yml files), not in the database (`activeModules` config option is completely removed)
+- Read module class extensions chain direct from the shop configuration (yml files). Don't store active module chain in the database (`aModules` config option is completely removed)
 - Don't store module settings in database
 - Change OXID eShop Community Edition license
 
@@ -51,6 +52,10 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - `OxidEsales\EshopCommunity\Core\Module\ModuleCache`
   - `OxidEsales\EshopCommunity\Core\Module\ModuleInstaller`
   - `Internal\Framework\Templating\Resolver\LegacyTemplateNameResolver`
+  - `OxidEsales\EshopCommunity\Internal\Framework\Module\Setup\Bridge\ClassExtensionChainBridgeInterface`
+  - `OxidEsales\EshopCommunity\Internal\Framework\Module\Setup\Bridge\ClassExtensionChainBridge`
+  - `OxidEsales\EshopCommunity\Internal\Framework\Module\Setup\Service\ClassExtensionChainServiceInterface`
+  - `OxidEsales\EshopCommunity\Internal\Framework\Module\Setup\Service\ClassExtensionChainService`
   - `OxidEsales\Eshop\Core\Routing\ModuleControllerMapProvider`
   - `OxidEsales\EshopCommunity\Internal\Framework\Module\Setup\Handler\ControllersModuleSettingHandler`
   - `OxidEsales\EshopCommunity\Internal\Framework\Module\Setting\SettingDao`
@@ -110,6 +115,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - `oxid_esales.module.setup.version_module_setting_handler`
   - `oxid_esales.module.setup.shop_configuration_class_extension_handler`
   - `oxid_esales.module.setup.events_module_setting_handler`
+  - `oxid_esales.module.setup.class_extension_chain_service`
   - `oxid_esales.module.setup.controllers_module_setting_handler`
   - `OxidEsales\EshopCommunity\Internal\Framework\Module\Setting\SettingDaoInterface`
   - `OxidEsales\EshopCommunity\Internal\Framework\Module\Setting\SettingDaoBridgeInterface`
@@ -121,6 +127,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
   - `aModuleEvents`
   - `aModuleControllers`
   - `activeModules`
+  - `aModules`
 - Constants:
   - `OxidEsales\EshopCommunity\Internal\Framework\Config\DataObject\ShopConfigurationSetting`
     - `MODULE_CLASS_EXTENSIONS`
@@ -129,6 +136,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
     - `MODULE_EVENTS`
     - `MODULE_CONTROLLERS`
     - `ACTIVE_MODULES`
+    - `MODULE_CLASS_EXTENSIONS_CHAIN`
   - `OxidEsales\EshopCommunity\Core\Module\ModuleList`
     - `MODULE_KEY_PATHS`
     - `MODULE_KEY_EVENTS`